pyproject.toml
[project]
name = "hgsp-lvm"
version = "0.1.0"
description = "Hilbert space Gaussian process latent variable model exploration in numpyro"
authors = [
    {name = "Brendan Cooley", email = "brendan.cooley@gmail.com"},
]
dependencies = [
    "modal~=0.63",
    "numpyro>=0.15.1,~=0.15",
    "optax~=0.2",
    "pydantic~=2.8",
    "python-dotenv~=1.0",
    "s3fs~=2024.1",
]
requires-python = "~=3.11"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.dev-dependencies]
lint = [
    "ruff>=0.5.2",
]

[tool.ruff.lint]
select = ["I"]